Zombie Clustering using python
There are zombies in Seattle. Liv and Ravi are trying to track them down to find out who is
creating new zombies in an effort to prevent an apocalypse. Other than the patient-zero
zombies, new people only become zombies after being scratched by an existing zombie.
Zombiism is transitive. This means that if zombie 0 knows zombie 1 and zombie 1 knows
zombie 2, then zombie 0 is connected to zombie 2 by way of knowing zombie 1. A zombie
cluster is a group of zombies who are directly or indirectly linked through the other zombies
they know, such as the one who scratched them or supplies who them with brains.
The diagram showing connectedness will be made up of a matrix of values 0 or 1. The value
matrix[i, j] indicates if the zombie i knows the zombie j. For example, in a world of 3 zombies the
complete matrix of zombie connectedness could be:
110
110
001
Zombies 0 and 1 know each other. Zombie 2 does not know anyone.